Move permission stuff to PermissionUtils
